The composer's contract moves out of the file that assembles it
`EditPraxisState` — the settled seam nine archetypes, the waiting surface,
the seal components and the dispatcher all read — was 200 lines of interface
sitting on top of its own implementation. It moves to `editPraxisState.ts`,
type-only so it costs no chunk bytes, and is re-exported from
`useEditPraxis.ts` so no importer changes.
